% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_ndvi.R
\name{diff_ndvi}
\alias{diff_ndvi}
\title{NDVI Difference Calculation for two timesteps}
\usage{
diff_ndvi(t1_nir, t1_red, t2_nir, t2_red, mask)
}
\arguments{
\item{t1_nir}{Near Infrared band for the first period}

\item{t1_red}{Red band for the first period}

\item{t2_nir}{Near Infrared band for the second period}

\item{t2_red}{Red band for the second period}

\item{mask}{A single raster layer containing only forested areas, used to mask out non-forested areas}
}
\value{
A single raster layer representing NDVI Difference between two timesteps
}
\description{
NDVI Difference Calculation for two timesteps
}
\examples{
library(forchange)
library(raster)
library(sp)
library(sf)
library(RStoolbox)
library(caret)
library(randomForest)
library(terra)


t1_nir_path <- system.file("extdata", "l5_nir.tif", package="forchange")
t1_nir <- raster(t1_nir_path)

t1_red_path <- system.file("extdata", "l5_red.tif", package="forchange")
t1_red <- raster(t1_red_path)


t2_nir_path <- system.file("extdata", "l8_nir.tif", package="forchange")
t2_nir <- raster(t2_nir_path)

t2_red_path <- system.file("extdata", "l8_red.tif", package="forchange")
t2_red <- raster(t2_red_path)

mask_path <- system.file("extdata", "forest_mask.tif", package="forchange")
mask<- raster(mask_path)



d_ndvi <- diff_ndvi(t1_nir, t1_red, t2_nir, t2_red, mask)
plot(d_ndvi)






}
