% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forchange.R
\name{forchange}
\alias{forchange}
\title{Decadal Forest Change Calculation}
\usage{
forchange(diff_NDVI, threshold)
}
\arguments{
\item{diff_NDVI}{A raster layer containing NDVI difference between two periods}

\item{threshold}{A specific value to be set to determine the extent of change}
}
\value{
A data frame containing forest change area and no change area
}
\description{
Decadal Forest Change Calculation
}
\examples{
library(forchange)
library(raster)
library(sp)
library(sf)
library(RStoolbox)
library(caret)
library(randomForest)
library(terra)





t1_nir_path <- system.file("extdata", "l5_nir.tif", package="forchange")
t1_nir <- raster(t1_nir_path)

t1_red_path <- system.file("extdata", "l5_red.tif", package="forchange")
t1_red <- raster(t1_red_path)


t2_nir_path <- system.file("extdata", "l8_nir.tif", package="forchange")
t2_nir <- raster(t2_nir_path)

t2_red_path <- system.file("extdata", "l8_red.tif", package="forchange")
t2_red <- raster(t2_red_path)

mask_path <- system.file("extdata", "forest_mask.tif", package="forchange")
mask<- raster(mask_path)

diff_ndvi(t1_nir, t1_red, t2_nir, t2_red, mask)

diff_NDVI <- diff_ndvi(t1_nir, t1_red, t2_nir, t2_red, mask)

threshold <- -0.2


forchange(diff_NDVI, threshold)




}
