% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formask.R
\name{formask}
\alias{formask}
\title{Forest Mask Calculation}
\usage{
formask(ras_stack, bandnames, tr_samp)
}
\arguments{
\item{ras_stack}{A raster layer stack  containing multiple bands}

\item{bandnames}{A character vector specifying the names of the bands in the raster layer stack}

\item{tr_samp}{A vector file containing training polygons for land-use classification, used to extract forest area}
}
\value{
A single raster layer representing  only the  forested areas within the specified region of interest
}
\description{
Forest Mask Calculation
}
\examples{
library(forchange)
library(raster)
library(sp)
library(sf)
library(RStoolbox)
library(caret)
library(randomForest)
library(terra)

ras_stack_path <- system.file("extdata", "roi_2000.tif", package="forchange")
ras_stack <- stack(ras_stack_path)

tr_samp_path <- system.file("extdata", "samples2000.gpkg", package = "forchange")
tr_samp <- st_read(tr_samp_path)

bandnames <- c("B01", "B02", "B03", "B04", "B05", "B06", "B07")

formask(ras_stack,bandnames,tr_samp)





}
